Extract HTML sidebar renderer
