+++ docs/DECISIONS.md
@@ -160,6 +160,7 @@
 - Static HTML browser rendering lives in `WorkspaceHTMLBrowserRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates browser pane, preview, snapshot, outline, text snippet, and comment markup while staying responsible for only the whole-workspace harness composition.
 - Static HTML secondary pane rendering lives in `WorkspaceHTMLSecondaryPaneRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates Extensions, Memories, Activity, and Automations pane markup while MCP extension chips, memory cards, activity sections, automation actions, and secondary count labels stay beside the secondary-pane HTML renderer.
 - Static HTML review rendering lives in `WorkspaceHTMLReviewRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates Git review pane, file, hunk, line, inline comment, and action markup while staying responsible for transcript placement.
+- Static HTML sidebar rendering lives in `WorkspaceHTMLSidebarRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates project rows, thread groups, bulk-selection controls, primary actions, and the tools/settings footer while staying responsible for whole-workspace shell composition.
 - Browser comment, snapshot, and browser panel state live in `QuillCodeBrowserSurface.swift`; browser address resolution lives in `WorkspaceBrowserLocationResolver.swift`; and browser history/page/comment state transitions live in `WorkspaceBrowserEngine`. The workspace model owns async fetch orchestration, `lastError`, and top-bar refreshes, while reusable browser state contracts, URL-resolution rules, and pure state transitions stay focused and guarded by parity tests.
 - Review comment payload state and event planning live in `WorkspaceReviewCommentPlanner.swift`, not in `WorkspaceModel.swift`. The planner owns path/text trimming, visible review-file validation, line-range normalization, line-kind checks, summary copy, and payload encoding. The workspace model owns only selected-thread validation, event append, persistence, and top-bar refresh.
 - Agent tool execution override composition lives in `WorkspaceToolExecutionOverrideCombiner.swift`, not in `WorkspaceModel.swift`. The combiner owns the precedence order Plan, Remote Project, Browser, Computer Use, Memory, MCP, while the workspace model only creates each optional executor and assigns the combined override to the runner.
